+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Wallender 2021 piperaquine PK, malaria-hazard and QTc models ([doi:10.1038/s41467-021-27051-8](https://doi.org/10.1038/s41467-021-27051-8)) -- Ugandan children 2 to 36 months of age receiving malaria chemoprevention.
+- Add Chen 2021 lorlatinib exposure-response models ([doi:10.1002/cpt.2228](https://doi.org/10.1002/cpt.2228)) -- adults with ALK-positive or ROS1-positive non-small cell lung cancer.
 
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.